don't store orientation (either 1 or 2) in a single bit... i check all other occurences off orientation and found no other such error
[git-p4: depot-paths = "//depot/qt/main/": change = 40221]

+++ src/widgets/qscrollbar.cpp
@@ -265,7 +265,7 @@ void QScrollBar::init()
 
 void QScrollBar::setOrientation( Orientation orientation )
 {
-    if ( (uint) orientation == orient )
+    if ( orientation == orient )
 	return;
 
     QSizePolicy pol = sizePolicy();
